requirements.txt
art==6.4
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
defusedxml==0.7.1
dnspython==2.7.0
idna==3.10
ipwhois==1.3.0
progress==1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.5
requests==2.32.3
six==1.17.0
soupsieve==2.6
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
validators==0.34.0
